Makes new STringExtensions file
